+++ MagnumProjekt/Nahrung/INahrungsmittel.cs
@@ -2,7 +2,7 @@ namespace MagnumProjekt;
 
 public interface INahrungsmittel 
 {
-    public int Brennwert { get; }
+    public double Brennwert { get; }
     public double Eiweiß { get; }
     public double Kohlenhydrat { get; }
     public double Fett { get; }
